M02A02 – Coerção de Tipo

+++ Modulo_02_Operadores/M02A02–Coercao_de_Tipo_02.c
@@ -8,7 +8,7 @@ void main(){
     printf("Digite um número: ");
     scanf("%d", &n);
 
-    float t = n/3;
+    float t = (float)n/3;
 
     printf("a terca parte de %d e igual a %.2f", n, t);
 }
